requirements.txt
aiofiles==24.1.0
aiosmtplib==2.0.2
aiosqlite==0.20.0
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
async-timeout==4.0.3
asyncpg==0.30.0
bcrypt==4.2.0
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
botocore==1.35.57
Brotli==1.1.0
brotli-asgi==1.4.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
cssselect==1.2.0
cssutils==2.11.1
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
execnet==2.1.1
fastapi==0.115.3
fastapi-cli==0.0.5
fastapi-mail==1.4.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.5.0
orjson==3.10.11
packaging==24.1
parso==0.8.4
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
premailer==3.10.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pylint==3.3.1
pyotp==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
requests==2.32.3
rich==13.9.3
rsa==4.9
s3transfer==0.10.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
sse-starlette==2.1.3
starlette==0.41.2
tenacity==9.0.0
tomli==2.0.2
tomlkit==0.13.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uuid6==2024.7.10
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
